refactor: split device_monitor into per-platform modules
